Clarify that nan() was a function, rename nan() to nan_dbl().
Before, there was a nan() function returning not-a-number.
This had to be defined as
    real(kind=dbl) :: nan
This very much looks like a (uninitialized) variable.

Now, I made nan.f90 to a module which has to be explicitly included using the USE-statement.
Also the nan() function is renamed do nan_dbl() to avoid a double name usage error.

Also:
Compile nan.f90 earlier to have it available as module.