readHIRHAM has been added to the importer if pyPamtra. This processes the HIRHAM Arctic output. It further includes the echam6 particle size distributions 'mono_echam_cl' and 'mon_echam_ice', that need to be defined in the descriptor file. It requires specific descriptor files.
Furthermore, minor changes in surface matrices for land surface, respectively the specular treatment have been added according to Xinxin Xie's suggestion.

A last change is due to the TELSEM2 emissivity maps. For some coastal points they returned 0 as emissivity. In this case we switch to the old SSMI based maps for this point.